Move files to lib/ directory
