+++ python/lib/sequence.py
